added kafka acknowledgement.

+++ src/main/java/com/maan/eway/notification/req/Sms.java
@@ -25,6 +25,6 @@ public class Sms implements Serializable {
 	private String whatsappRegards;
 	private JobCredentials credential;
 	private String senderId;
-	  private Integer    notifNo ;
+	private Long    notifNo ;
 
 }
